chore(containers): reorganise exports to prepare 6.0
