+++ packages/containers/6.0 - breaking changes.md
@@ -1,0 +1,69 @@
+# 6.0 Breaking changes
+
+## AboutDialog
+
+Constants are now attached to the component.
+
+```diff
+-import AboutDialogConstants from '@talend/react-containers/lib/AboutDialog/AboutDialog.constants'
++import AboutDialog from '@talend/react-containers/lib/AboutDialog';
+
++const AboutDialogConstants = AboutDialog.constants;
+```
+
+## EditableText
+
+Selectors are now attached to the component
+
+```diff
+-import { getEditMode } from '@talend/react-containers/lib/EditableText/EditableText.selectors'
++import EditableText from '@talend/react-containers/lib/EditableText';
+
++const { getEditMode } = EditableText.selectors;
+```
+
+## GuidedTour
+
+Constants are now attached to the component.
+
+```diff
+-import GuidedTourConstants from '@talend/react-containers/lib/GuidedTour/GuidedTour.constants'
++import GuidedTour from '@talend/react-containers/lib/GuidedTour';
+
++const GuidedTourConstants = GuidedTour.constants;
+```
+
+## HeaderBar
+
+Constants and actions are now attached to the component.
+
+```diff
+-import HeaderBarConstants from '@talend/react-containers/lib/HeaderBar/HeaderBar.constants'
+-import HeaderBarActions from '@talend/react-containers/lib/HeaderBar/HeaderBar.actions'
++import HeaderBar from '@talend/react-containers/lib/HeaderBar';
+
++const HeaderBarConstants = HeaderBar.constants;
++const HeaderBarActions = HeaderBar.actions;
+```
+
+## List
+
+Constants and selectors are now attached to the component.
+
+```diff
+-import ListConstants from '@talend/react-containers/lib/List/List.constants'
+-import { getCollectionItems } from '@talend/react-containers/lib/List/selector'
++import List from '@talend/react-containers/lib/List';
+
++const ListConstants = List.constants;
++const { getCollectionItems } = List.selectors;
+```
+
+## registerAllContainers
+
+`registerAllContainers` function is now exported from index
+
+```diff
+-import { registerAllContainers } from '@talend/react-containers/lib/register'
++import { registerAllContainers } from '@talend/react-containers';
+```
