+++ packages/containers/src/ActionFile/ActionFile.connect.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cmf, { cmfConnect } from '@talend/react-cmf';
-import ActionFile from '@talend/react-components/lib/Actions/ActionFile';
+import { ActionFile } from '@talend/react-components/lib/Actions';
 
 export function mapStateToProps(state, ownProps) {
 	if (!ownProps.actionId) {
