+++ packages/containers/src/ActionSplitDropdown/ActionSplitDropdown.connect.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cmf, { cmfConnect } from '@talend/react-cmf';
-import ActionSplitDropdown from '@talend/react-components/lib/Actions/ActionSplitDropdown';
+import { ActionSplitDropdown } from '@talend/react-components/lib/Actions';
 
 import getOnClick from '../actionOnClick';
