+++ packages/containers/src/EditableText/index.js
@@ -1,3 +1,5 @@
 import EditableText from './EditableText.connect';
+import * as EditableTextSelectors from './EditableText.selectors';
 
+EditableText.selectors = EditableTextSelectors;
 export default EditableText;
