+++ packages/containers/src/GuidedTour/index.js
@@ -1,7 +1,9 @@
 import GuidedTour from './GuidedTour.connect';
 
 import GuidedTourSagas from './GuidedTour.sagas';
+import GuidedTourConstants from './GuidedTour.constants';
 
 GuidedTour.sagas = GuidedTourSagas;
+GuidedTour.constants = GuidedTourConstants;
 
 export default GuidedTour;
