+++ packages/containers/src/register.js
@@ -1,7 +1,7 @@
 import cmf from '@talend/react-cmf';
-import containersCMFModule from './index';
+import cmfModule from './cmfModule';
 
 // eslint-disable-next-line import/prefer-default-export
 export function registerAllContainers() {
-	cmf.component.registerMany(containersCMFModule.components);
+	cmf.component.registerMany(cmfModule.components);
 }
